New function defining syntax
Now functions won't require to be defined as `f = sign "", f = f ->`.
This was because the call to `f` after being initialized with a
signature created a new object. Now, the function metatable changes
itself, therefore merging the "signed_constructor" and "signed" types
(although they still exist, "signed_constructor" is the type before
assigning a function, and it becomes "signed" just after). For
compatibility, you can access the old function by requiring
"ltypekit.wrap", but the new function is defined as `sign` in
"ltypekit.sign". requiring only "typekit" will import a table with the
`sign` function and an alias to it named `signature`.